refactor: Relocate UI mappers to dedicated package
