+++ presentation/src/commonMain/kotlin/com/moneyplusplus/presentation/mapper/TransactionTypeFilterMapper.kt
@@ -1,0 +1,10 @@
+package com.moneyplusplus.presentation.mapper
+
+import com.moneyplusplus.domain.entity.TransactionType
+import com.moneyplusplus.presentation.feature.transaction.TransactionUiState
+
+fun TransactionUiState.TransactionTypeFilter.toDomain() = when (this) {
+    TransactionUiState.TransactionTypeFilter.ALL -> null
+    TransactionUiState.TransactionTypeFilter.INCOMES -> TransactionType.INCOME
+    TransactionUiState.TransactionTypeFilter.EXPENSES -> TransactionType.EXPENSE
+}
