+++ presentation/src/commonMain/kotlin/com/moneyplusplus/presentation/model/TransactionUiModel.kt
@@ -1,10 +1,5 @@
 package com.moneyplusplus.presentation.model
 
-import com.moneyplusplus.domain.entity.Transaction
-import com.moneyplusplus.domain.entity.TransactionType
-import com.moneyplusplus.presentation.formater.formatAmount
-import com.moneyplusplus.presentation.formater.formatDate
-
 data class TransactionUiModel(
     val id: String,
     val title: String,
@@ -14,12 +9,3 @@ data class TransactionUiModel(
     val categoryId: String
 )
 
-fun Transaction.toUiModel() =
-    TransactionUiModel(
-        id = id.toString(),
-        title = title,
-        amount = "${amount.formatAmount()} ${currency.code}",
-        isExpense = type == TransactionType.EXPENSE,
-        dateFormated = date.formatDate(),
-        categoryId = category.id.toString()
-    )
